Update netty 4 common package name
