+++ instrumentation/netty/netty-4-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/netty/v4/common/client/NettyErrorOnlyConnectionInstrumenter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.netty.common.client;
+package io.opentelemetry.javaagent.instrumentation.netty.v4.common.client;
 
 import io.netty.channel.Channel;
 import io.opentelemetry.context.Context;
@@ -11,27 +11,31 @@ import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
 import io.opentelemetry.javaagent.instrumentation.netty.common.NettyConnectionRequest;
 import javax.annotation.Nullable;
 
-final class NettyConnectionInstrumenterImpl implements NettyConnectionInstrumenter {
+final class NettyErrorOnlyConnectionInstrumenter implements NettyConnectionInstrumenter {
 
   private final Instrumenter<NettyConnectionRequest, Channel> instrumenter;
 
-  NettyConnectionInstrumenterImpl(Instrumenter<NettyConnectionRequest, Channel> instrumenter) {
+  NettyErrorOnlyConnectionInstrumenter(Instrumenter<NettyConnectionRequest, Channel> instrumenter) {
     this.instrumenter = instrumenter;
   }
 
   @Override
   public boolean shouldStart(Context parentContext, NettyConnectionRequest request) {
-    return instrumenter.shouldStart(parentContext, request);
+    // the "real" check is done on end() anyway
+    return true;
   }
 
   @Override
   public Context start(Context parentContext, NettyConnectionRequest request) {
-    return instrumenter.start(parentContext, request);
+    return parentContext;
   }
 
   @Override
   public void end(
       Context context, NettyConnectionRequest request, Channel channel, @Nullable Throwable error) {
-    instrumenter.end(context, request, channel, error);
+    if (error != null && instrumenter.shouldStart(context, request)) {
+      Context connectContext = instrumenter.start(context, request);
+      instrumenter.end(connectContext, request, channel, error);
+    }
   }
 }
