+++ instrumentation/netty/netty-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/netty/v4_0/NettyInstrumentationModule.java
@@ -12,7 +12,7 @@ import static net.bytebuddy.matcher.ElementMatchers.not;
 import com.google.auto.service.AutoService;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.instrumentation.netty.common.NettyFutureInstrumentation;
+import io.opentelemetry.javaagent.instrumentation.netty.v4.common.NettyFutureInstrumentation;
 import java.util.List;
 import net.bytebuddy.matcher.ElementMatcher;
